Keyedin/requirements.txt
# KeyedIn Enhanced API Requirements

# Core dependencies
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Environment variable management
python-dotenv>=1.0.0

# Optional but recommended
beautifulsoup4>=4.12.0  # For HTML parsing if needed
lxml>=4.9.0  # For faster HTML parsing


